NOW COMPATIBLE WITH 1.9.2+svn HECK YEAH! Lots and lots of bug fixes. Teleregeneration should now work for steelhive. Many linkages patched up.
I inadvertently added alot of bugs but they are all now fixed, as if a special note definition's #enddef does not end on the same line as the last ", multiple such macros in a row cause errors. These can be spotted by "unexpected start of line symbol +". All those are fixed now.