AtS E2S4: slight tweaks to the initial cutscene to prevent faerie halo glitches
Haloes are (re)moved from the map when [store_unit] is used, but not
when [teleport] is used, thus leaving Anya and Tara's halos behind until
they are finally stored away at the end of the cutscene.

This commit changes the cutscene such that Anya and Tara are stored and
removed from the map along with their halos immediately after they
teleport away.