AtS: "remove" Elyssa's guardian-with-a-sword variation's sword magical weapon special
The magical weapon special is inherited from the "claw of urvatha"
weapon due to the [variation] inheritance mechanism. Since the sword
attack already has drains, this breaks weapon special conventions.

The solution is making the special both invisible in the UI and cause it
to filter on an impossible condition.