AtS: globally hook into preload events to check for Wesnoth version compatibility
Wesnoth versions < 1.9.10 / < 1.10, and versions 1.11.0 and 1.11.1 are
not supported at this time due to known confirmed or potential bugs. In
order to avoid spurious bug reports in the future it's best to prevent
the player from trying to play on those versions to some degree. The
warning message is still dimissable in case the player chooses to not
heed my perfectly reasonable advice.