AtS: new improved workaround for bogus "should not have a translation mark" wmllint warnings
These warnings originate from untranslatable Pango markup and
translatable string concatenation using the parser + token.

For example:

  "<small><span color='#f00' weight='bold'>!</span> <span color='#888'>"+_"Music add-on “<i>AtS Music</i>” is <b>not</b> currently installed."+"</span></small>"

When inserting an unquoted newline immediately after the + token, the
parser will not include it in the resultant string, and wmllint won't
misparse the line as a misapplied translatable string attribute:

  "<small><span color='#f00' weight='bold'>!</span> <span color='#888'>"+
  _"Music add-on “<i>AtS Music</i>” is <b>not</b> currently installed."+"</span></small>"

This solution should prove sufficient for the time being. Tested both on
Wesnoth 1.10 and 1.11.6+dev.